refactor: inline clash manager logic
